Move font parsing into asset uploader
Here I moved font parsing into asset uploader as it is used only there
but bloats sdk and cli with big fontkit package.

Cli install size

```diff
du -ck node_modules
- 72464
+ 57880
```
