+++ packages/fonts/src/constants.ts
@@ -19,3 +19,6 @@ export const FONT_FORMATS: Map<FontFormat, string> = new Map([
 export const FONT_MIME_TYPES = Array.from(FONT_FORMATS.keys())
   .map((format) => `.${format}`)
   .join(", ");
+
+export const FONT_STYLES = ["normal", "italic", "oblique"] as const;
+export type FontStyle = (typeof FONT_STYLES)[number];
