--- packages/fonts/src/index.server.ts
@@ -1,1 +1,0 @@
-export { getFontData } from "./font-data";
